pyproject.toml
[project]
name = "langprobe"
version = "0.1.0"
description = "DSPy benchmarking suite"
authors = []
dependencies = [
    "requests>=2.32.3",
    "shortuuid",
    "seaborn",
    "black==23.11.0",
    "sentence-transformers==2.2.2",
    "sentencepiece==0.1.99",
    "huggingface-hub==0.25.0",
    "numpy==1.24.3",
    "torch>=2.2.2",
    "python-dotenv>=1.0.1",
    "ipython>=8.30.0",
    "dspy @ git+https://github.com/stanfordnlp/dspy.git@miprov2-knn",
    "urllib3>=2.2.3",
    "alfworld>=0.4.1",
]
readme = "README.md"
requires-python = ">=3.11,<3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["langProBe"]

[tool.ruff]
line-length = 120
